pyproject.toml
[project]
name = "factorio-bp-editor"
version = "0.1.0"
description = "A modern Python tool for editing and manipulating Factorio blueprints"
readme = "README.md"
authors = [
    { name = "Chris Busillo", email = "info@shinycomputers.com" }
]
requires-python = ">=3.13"
dependencies = [
    "factorio-draftsman>=2.0.2",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-cov>=5.0.0",
    "ruff>=0.8.0",
    "mypy>=1.13.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]

[tool.ruff]
line-length = 100
target-version = "py313"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W"]
ignore = ["E501"]

[tool.mypy]
python_version = "3.13"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true

[[tool.mypy.overrides]]
module = "draftsman.*"
ignore_missing_imports = true
